Improve content negotiation API
Previously, the API require user to pass list of tuples (media_type,
unit -> string). This was not very ergonomic, and there is no way to
express 406 not acceptable response.

Now, the API instead require a function of format -> (string * string)
option. When the function return None, the server will respond with
406 Not Acceptable. If it return Some (media_type, body), the server
will respond normally with status code given to Response.negotiate
function.