feature(router): prepare for OpenAPI generation
In preparation of OpenAPI schema generation, we need to refactor our router definition.
The problem of previous implementation is we handle many things in path pattern.
As a result, the (/) combinator can fail at runtime, and it hard to
extract the pattern. Also we missing the parameter name data.

now we separate path pattern and schema. So we can easily extract the path patterns,
the schema handle the high level logic like request body, method, etc.

This PR is big since the router is the heart of the framework.